Bug #1392309 cpufreq: sprd: fix Coverity warning
[root cause  ] Coverity warning
[changes     ] fix Coverity warning
[side effects] none
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: none
[confirm dependent commit]: none
[board]: s9863a-1h10
[test case]: dvfs
[reviewers   ] luyao.wu jack.liu
[change_type ] feature_bugfix
[tag_product ] common

Fix Coverity warnings:
	Modify the variable type, and the local variable
	has no practical meaning in the function

+++ drivers/cpufreq/sprd-cpufreqhw.c
@@ -270,7 +270,7 @@ static int sprd_hardware_cpufreq_init(struct cpufreq_policy *policy)
 	struct sprd_cpudvfs_device *pdev;
 	struct sprd_cpudvfs_ops *driver;
 	struct device_node *cpu_np;
-	unsigned long freq_Hz = 0;
+	unsigned int freq_Hz = 0;
 	struct device *cpu_dev;
 	int ret, cpu = 0;
 	int curr_cluster;
